+++ Runtime/Unity/BehaviourExtensions.cs
@@ -1,0 +1,22 @@
+﻿using UnityEngine;
+
+namespace Mirzipan.Extensions.Unity
+{
+    public static class BehaviourExtensions
+    {
+        public static void Enable(this Behaviour @this)
+        {
+            @this.enabled = true;
+        }
+        
+        public static void Disable(this Behaviour @this)
+        {
+            @this.enabled = false;
+        }
+
+        public static bool IsPrefab(this Behaviour @this)
+        {
+            return @this.gameObject.IsPrefab();
+        }
+    }
+}