+++ Runtime/Unity/ComponentExtensions.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace Mirzipan.Extensions.Unity
 {
@@ -12,6 +13,16 @@ namespace Mirzipan.Extensions.Unity
             }
             
             @this.gameObject.SetActive(value);
+        }
+        
+        public static T GetOrAddComponent<T>(this Component @this) where T: Component
+        {
+            return @this.GetComponent<T>() ?? @this.gameObject.AddComponent<T>();
+        }
+
+        public static Component GetOrAddComponent(this Component @this, Type type)
+        {
+            return @this.GetComponent(type) ?? @this.gameObject.AddComponent(type);
         }
     }
 }